requirements.txt
Flask==2.1.2
spotipy~=2.19.0
python-dotenv~=0.20.0